Enables SPA routing for GitHub Pages deployment
Configures the application to support Single Page App (SPA) routing when deployed on GitHub Pages.

This commit removes the `basename` property from the `Router` component in `App.jsx` and updates the `base` property in `vite.config.js` to `/`.
It also adds a 404.html in the public folder and a script to the index.html.
These changes ensure that client-side routing works correctly in a GitHub Pages environment.